modify and reuse exisiting methods
